Crate glsl_lang_lexer

Modules

context 🔒

full

glsl-lang-pp/full based lexer

lang_token 🔒

min

glsl-lang-pp/min based lexer

token 🔒

Structs

GlslTypeTablePolicy

Default GLSL type table policy: only struct declarations create new type names

ParseContext

Parsing context

ParseContextData

Parsing context data

ParseOptions

Parsing options

Enums

IdentifierContext

Context in which an identifier is seen for the first time

Token

Traits

HasLexerError

Language lexer error definition

LangLexer

GLSL language lexer

LangLexerIterator

GLSL language lexer iterator

TypeTablePolicy

A policy to dictate which identifiers should be seen as type names or not

Type Aliases

CommentList

A list of comments indexed by their position