Struct glsl_lang_lexer::context::ParseContextData

pub struct ParseContextData {
    names: Vec<HashSet<SmolStr>>,
    comments: Option<CommentList>,
    policy: Rc<dyn TypeTablePolicy>,
}

Parsing context data

Fields
names: Vec<HashSet<SmolStr>>

List of known type names

comments: Option<CommentList>

List of parsed comments (or None to disable comment parsing)

policy: Rc<dyn TypeTablePolicy>
Implementations

impl ParseContextData

pub fn new() -> Self

Create a new ParseContextData object

pub fn with_policy(policy: impl TypeTablePolicy + 'static) -> Self

Create a new ParseContextData object with the given type table policy

Parameters

policy: policy object managing the type table

pub fn with_comments() -> Self

Create a new ParseContextData object with comments parsing enabled

pub fn with_comments_and_policy(policy: impl TypeTablePolicy + 'static) -> Self

Create a new ParseContextData object with comments parsing enabled and the given type table policy

Parameters

policy: policy object managing the type table

pub fn comments(&self) -> Option<&CommentList>

Get the list of comments stored in this parse context

impl ParseContextData

pub fn is_type_name(&self, name: &str) -> bool

Return true if the given name is a type name

pub fn add_type_name(&mut self, name: Identifier) -> TypeName

Register name as a new type name

pub fn push_scope(&mut self)

Enter a new nesting level for declarations

pub fn pop_scope(&mut self)

Leave the current nesting level

pub fn new_identifier(&mut self, name: &Identifier, ctx: IdentifierContext)

Update the context data with a new identifier in a given context

impl ParseContextData

pub fn add_comment(&mut self, comment: Comment)

Add a new comment to the parsed comments list