Trait glsl_lang_lexer::context::TypeTablePolicy

pub trait TypeTablePolicy: Debug {
    fn promote_to_type_name(
        &self,
        name: &Identifier,
        ctx: IdentifierContext
    ) -> bool;
}

A policy to dictate which identifiers should be seen as type names or not

Required Methods

fn promote_to_type_name(&self, name: &Identifier, ctx: IdentifierContext) -> bool

Return true if the given identifier (in its context) should now be considered a type name or not.